Revert "adding clip 0, 1 to all eps bound calculations"
This reverts commit 4bea78deeedc1f1c297dba4387402729bd2f4df1.

+++ script/settings.py
@@ -1,4 +1,4 @@
 import torch
 
 device = "cpu"
-data_type = torch.float32
+data_type = torch.float64
